Enable return in textfield to submit the newsletter
I have a personal preference to __never__ use my mouse unless I can handle it.
It's very likely that my browser would allow me to tab from the textfield to
the button to submit, but I'm lazy and greatly prefer to use return to submit
forms.

This commit adds a form to the page, and abstracts out the click handler on the input button and uses it as the submit handler on the form. It returns false
to avoid falling back to the default form submit, which isn't what you want to
do here at all.

The button should probably be a submit thing that makes the form post, for extra
bonus points, but this is functional.